feat(engine): support Transferable RPC payloads
- Extend EngineWorker.invoke() with an optional transfer list and forward it to postMessage
- Add loadWorkbookFromXlsxBytes(Uint8Array) API that transfers bytes.buffer to the worker
- Wire worker handler for loadFromXlsxBytes and document transfer/detach semantics
- Update EngineWorker tests to assert transfer list usage

+++ packages/engine/src/engine.worker.ts
@@ -231,7 +231,7 @@ async function handleRequest(message: WorkerInboundMessage): Promise<void> {
         break;
       case "loadFromXlsxBytes":
         {
-          const next = mod.WasmWorkbook.fromXlsxBytes(params.bytes);
+          const next = mod.WasmWorkbook.fromXlsxBytes(params.bytes as Uint8Array);
           freeWorkbook(workbook);
           workbook = next;
         }
